Dropping the devise and clearance in favor of simple auth
Trying to bring it back to something a little more simplistic that I can understand.  In the future, I'm going to have to be able to really understand this stuff cuz it's going to be needed no matter what.

But...for now.  This should work and it'll teach me to get it right the first time.