Clean up, redirection, adding clubs
Dealt with some leftover stuff from the devise days.  Added in the clubs and took the interesting approach of adding a piece at a time.  While I really want to switch to a Cucumber/RSpec approach, I don't have a lot of time for it now.  I simply want to get this running first.

Next steps - get the rest of the models and forms in place and then think about styling this into something more useful.

Then move to Heroku for use at club.