Added blueprint, and worked with authentication ideas
For a while there I thought it would be a good idea to have members authenticate themselves. In the long run, I feel like this is a bad idea for a number of reasons.

The user model needs to be reimplimented.  Users are switched to identification by e-mail address as their username.  Ultimately, each user's e-mail address will be associated to many clubs.  While it's unlikely that someone will want to administer more than one club, it's possible, and that's where we'll have to have some sort of user's club listing.  If they only have one, then bop them directly over to the club of their choice.  Furthermore, a user will have to be able to be invited to administer another club online.

I'm still struggling with the path issues when pointing to the login on Sessions.  Thought I had that one nailed down, and now the nails have come loose.

Sigh.