Removed user, admin cruft.  Pulled club/member flow together
Got rid of some junk and moved into the archive drawer.  Clean is good.  Albeit not 'properly', added connections and links between the users and clubs.  Got the junky create member out of the club proper, need to much more clearly define the difference between a member and a club.  Colors, anyone?