Still learning to kick off plucky queries in MongoMapper
Fixed a bug with Members showing roles.
TODO: Understand point in time querying - split member's roles into history and upcoming.  The more I see of the Upcoming and Members the less I like it floating around on the left.  This will be particularly difficult with specific members and meeting editing. Already it's hard to understand.  Making this change is almost or perhaps harder than getting the code to work right.