Meeting form now in proper semantic form for.
TODO: Dates are formatted fully - On the one hand, it'd be really nice to use the actual time in the date, on the other that's a bit of a UX nightmare.  Maybe concatenate them? Use an attribute.  Research the semantic_form_for and see if there's a ready made - otherwise, this might be a pain.