Baisc structure in place for relationships.  Now invite members?
Inviting membership is something that's done all over the place.  I'm doing it badly because I only have a few members, but this will _not_ scale.

So:  How to search for, add exisiting members, and create new members if they're not already in the system?  How to reconcile the unique identity problem for members?  If they belong to a specific club... if they belong to a specific club.  If they belong to a specific club, yes that makes listing all memberships a little problematic (not really), but then I stop worrying about the edge cases of multi-club members and focus purely on getting it to work in a sensible manner.  All of the cludge that I've been fighting (and learning) through is immaterial. Yippee!